+++ docs/src/release_notes.rst
@@ -8,6 +8,15 @@
 Release Notes
 =============
 
+Upcoming
+--------
+
+* Wallet automation
+
+  * Fix an issue in the wallet sweep automation where it created
+    additional transfer offers even if there were already sufficient
+    transfer offers to cover the sweep.
+
 0.1.16
 ------